// frontend/src/background/interfaces.ts
export interface Playlist {
	/** Id of the playlist. */
	id: string
	/** Name of the playlist */
	name: string
}
// ...
export enum RemoteOperationType {
	Create = 'create',
	Read = 'read',
	Update = 'update',
	Delete = 'delete'
}

export interface RemoteOperation {
	type: RemoteOperationType
	// eslint-disable-next-line @typescript-eslint/no-explicit-any
	payload: any
}